src/backend/requirements.txt
Flask>=3.0
gunicorn>=22.0
requests>=2.31
python-cas>=1.6.0